refactor into separate confirm icons
