+++ src/components/job-definition-row.tsx
@@ -15,7 +15,7 @@ import TableCell from '@mui/material/TableCell';
 import { Scheduler, SchedulerService } from '../handler';
 import { useTranslator } from '../hooks';
 import { TranslationBundle } from '@jupyterlab/translation';
-import { ConfirmDeleteIcon } from './confirm-delete-icon';
+import { ConfirmDeleteIcon } from './confirm-icons';
 
 function CreatedAt(props: {
   job: Scheduler.IDescribeJobDefinition;
